Replace ReduxAuthTimeoutPopup with Login/AuthPopup
- does the same thing, but with useSelector & more compact
